+++ lib/skylight/messages/span.rb
@@ -21,7 +21,6 @@ module Skylight
 
         attr_reader \
           :time,
-          :token,
           :category,
           :title,
           :started_at,
@@ -30,9 +29,8 @@ module Skylight
 
         attr_accessor :children
 
-        def initialize(trace, time, token, started_at, cat, title, desc, annot)
+        def initialize(trace, time, started_at, cat, title, desc, annot)
           @trace      = trace
-          @token      = token
           @built      = false
           @time       = time
           @started_at = started_at
@@ -44,6 +42,10 @@ module Skylight
 
         def endpoint=(name)
           @trace.endpoint = name
+        end
+
+        def done
+          @trace.done(self) unless built?
         end
 
         def built?