+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## unreleased ##
 
 * [BUG] Fix unix domain socket write function in standalone agent
+* Performance improvements
 
 ## 0.1.5 (May 31, 2013)