+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * [BUG] Fix unix domain socket write function in standalone agent
 * Performance improvements
+* Tolerate invalid trace building
 
 ## 0.1.5 (May 31, 2013)